pyproject.toml
[project]
name = "evaluation"
version = "0.1.0"
description = "数据采集评估工具"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dash>=2.18.0",
    "dash-bootstrap-components>=1.6.0",
    "plotly>=5.24.0",
    "pandas>=2.2.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "minio>=7.2.0",
    "opencv-python-headless>=4.10.0",
    "Pillow>=10.0.0",
    "pyarrow>=15.0.0",
    "h5py>=3.11.0",
    "numpy>=1.26.0",
    "rq>=2.7.0",
    "redis>=7.2.1",
    "pyyaml>=6.0.3",
    "loguru>=0.7.3",
    "openai>=1.57.0",
]

[project.scripts]
start = "src.app:app.run"
